Add files via upload
The network reports on the inventors that are shared by different applicants. Inventors moving from one company to another or in a joint cooperation are taking their know-how with them, and so, the inventors in this graphic are the vectors between companies, indicating how the know-how flows between companies.
For the edges containing information of from where to where the know-how flows the priority date of the patents should be taken into account. That hasn't been done in this case, because, to a great extent the origins and destinies are quite intuitive. If a small company or a University shares know-how with a multinational company, it is quite clear in which direction the know-how flows.